+++ changelog.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 2025/08/02 08:29:10 Error starting transcription job: operation error Transcribe: StartTranscriptionJob, https response error StatusCode: 400, RequestID: 0fb0a4f6-5163-4cdb-8c5d-44491af38dbd, BadRequestException: The specified S3 bucket isn't in the same region. Make sure the bucket is in the eu-central-1 region and try your request again.
 
+## [v0.2.1]
+
+## Changed
+- Config files are now in ~/.config/audionote
+- If empty, embedded prompts are copied
 
 ## [v0.2.0]